requirements.txt
streamlit==1.35.0
scikit-image==0.18.3
opencv-python-headless==4.5.3.56
pillow==8.4.0
numpy==1.21.4
